refactor: modal-base, modal-button 분리 (#25)
